fix checkSession in AuthContext removing token parameter in getAuthUser function (don't necessary) + move useAuth to hooks dir + move AuthContextType interface to auth.ts in types dir
