+++ frontend/src/hooks/useAuth.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { AuthContext } from "./AuthContext";
+import { AuthContext } from "../contexts/AuthContext/AuthContext";
 
 export const useAuth = () => {
   const context = useContext(AuthContext);
