+++ frontend/src/types/auth.ts
@@ -12,3 +12,11 @@ export interface RegisterData {
     password: string;
     password_confirmation: string;
 }
+
+export interface AuthContextType {
+  user: User | null;
+  loading: boolean;
+  login: (email: string, password: string) => Promise<void>;
+  register: (userData: any) => Promise<void>;
+  logout: () => Promise<void>;
+}
